Player.java , Computer.java 생성. Application.java에서 빼내와 따로 분류해두었습니다.
